gpio: Split embedded-hal from main implementation
